+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,12 +1,15 @@
 package ru.yandex.practicum.filmorate.controller;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
 import java.util.List;
 
+@Slf4j
 @RestController
 @RequestMapping("/users")
 public class UserController {
@@ -18,12 +21,16 @@ public class UserController {
     }
 
     @GetMapping
-    public List<User> findAll() {
-        return userService.findAll();
+    public List<User> getAll() {
+        return userService.getAll();
     }
 
     @GetMapping("/{id}")
     public User getUser(@PathVariable Long id) {
+        if (!userService.getInMemoryUserStorage().getMap().containsKey(id)) {
+            log.error("При попытке получить данные пользователя возникла ошибка");
+            throw new NotFoundException("Пользователя с данным id не существует");
+        }
         return userService.getUser(id);
     }
 
@@ -34,11 +41,19 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable Long id) {
+        if (!userService.getInMemoryUserStorage().getMap().containsKey(id)) {
+            log.error("При попытке удалить пользователя возникла ошибка");
+            throw new NotFoundException("Пользователя с данным id не существует");
+        }
         userService.deleteUser(id);
     }
 
     @PutMapping
     public User update(@RequestBody User user) {
+        if (!userService.getInMemoryUserStorage().getMap().containsKey(user.getId())) {
+            log.error("При попытке обновить данные пользователя возникла ошибка");
+            throw new NotFoundException("Пользователя с данным id не существует");
+        }
         return userService.updateUser(user);
     }
